Design

+++ src/design/Employee.java
@@ -10,6 +10,7 @@ public interface Employee {
 	
 	//employeeId() will return employee id.
 	public int employeeId();
+
 	
 	//employeeName() will return employee name
 	public String employeeName();
@@ -18,7 +19,7 @@ public interface Employee {
     public void assignDepartment();
 	
 	//calculate employee salary
-	public int calculateSalary();
+	public double calculateSalary();
 	
 	//employee benefit
 	public void benefitLayout();
